refactor : 홈 Response에 UserLog 추가
