+++ src/main/java/com/coniverse/dangjang/domain/guide/bloodsugar/dto/SubGuideResponse.java
@@ -3,7 +3,7 @@ package com.coniverse.dangjang.domain.guide.bloodsugar.dto;
 import java.util.List;
 
 import com.coniverse.dangjang.domain.guide.bloodsugar.document.TodayGuide;
-import com.coniverse.dangjang.domain.guide.common.dto.GuideResponse;
+import com.coniverse.dangjang.domain.guide.common.dto.response.GuideResponse;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 /**
