+++ src/main/java/com/coniverse/dangjang/domain/guide/common/dto/response/DayGuideResponse.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang.domain.guide.common.dto;
+package com.coniverse.dangjang.domain.guide.common.dto.response;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -6,6 +6,7 @@ import java.util.List;
 import com.coniverse.dangjang.domain.guide.bloodsugar.document.TodayGuide;
 import com.coniverse.dangjang.domain.guide.exercise.dto.ExerciseDayGuide;
 import com.coniverse.dangjang.domain.guide.weight.dto.WeightDayGuide;
+import com.coniverse.dangjang.domain.log.dto.request.UserLog;
 
 /**
  * 하루 가이드 조회 응답
@@ -14,6 +15,6 @@ import com.coniverse.dangjang.domain.guide.weight.dto.WeightDayGuide;
  * @since 1.0.0
  */
 public record DayGuideResponse(String nickname, LocalDate date, List<TodayGuide> bloodSugars, WeightDayGuide weight, ExerciseDayGuide exercise,
-							   boolean notification) {
+							   boolean notification, UserLog userLog) {
 
 }
