+++ src/main/java/com/coniverse/dangjang/domain/guide/common/service/GuideService.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 import com.coniverse.dangjang.domain.analysis.dto.AnalysisData;
 import com.coniverse.dangjang.domain.code.enums.CommonCode;
 import com.coniverse.dangjang.domain.code.enums.GroupCode;
-import com.coniverse.dangjang.domain.guide.common.dto.GuideResponse;
+import com.coniverse.dangjang.domain.guide.common.dto.response.GuideResponse;
 
 /**
  * 가이드 service
