+++ src/main/java/com/coniverse/dangjang/domain/healthmetric/dto/response/HealthMetricResponse.java
@@ -2,7 +2,7 @@ package com.coniverse.dangjang.domain.healthmetric.dto.response;
 
 import java.time.LocalDate;
 
-import com.coniverse.dangjang.domain.guide.common.dto.GuideResponse;
+import com.coniverse.dangjang.domain.guide.common.dto.response.GuideResponse;
 
 /**
  * 건강지표 response dto이다.
